Add experimental support for GL OPTIC TOUCH device.
In order to add support for GL OPTIC device which communicates over TCP a more abstract BaseDevice was created as a simple skeleton. GlOpticTouch class directly inherits from this BaseDevice while all previously implemented serial classes use intermediate SerialDevice parent which further refines the use toward serial interface specifically. This change is in direction of scaling the package for support of devices with varying communication interfaces, as was done now with GL TOUCH support.

There are a few minor changes in the code along this commit:
- some serial classes did not expand/modify the parent __init__, yet they needlessly called super().__init__() on __init__. These lines of code could be deleted as parent __init__ is called on creation of a child instance if child does not attempt to overwrite it anyhow.
- some methods internal to get_, set_, etc have now a leading underscore to signify internal use intention. (e.g. read() -> _read()).
- add doc string to _arg_chech() method just to realize its terrible and too dynamic and causing plenty of more or less soft warnings. It appears as a very bad idea that _arg_check() always turns input into a tuple. It causes much of the downstream code to be boiler plate, difficult to read and leaves potential for some unchecked types to sit inside the tuple that will lead to errors that will be potentially difficult to trace to the root cause. _arg_check my dear you have to change for the better!