Improve _channel_arg_check method and its use to ensure transparent error handling on incorrect channel argument type or value in consumer methods.
In case user specifies other than permitted type for channel he should receive a clear traceback with root cause.
It is now no longer possible to set_output on multiple channels in one command using tuple for channels. This should instead always be done with a loop on consumer side. This change is intended to remove ambiguity in channels type and in order not to suggest that set_output can indeed be commenced in parallel on multiple inputs, which is not true for all devices implemented this far.