Prevent clipping of last character in return message in Tti3ChPsu class get_intput() method.
Ever since write and read terminations are allowed to be different (separate dict fields in DEFAULTS) the workaround to clip CR due to read termination being longer than write termination should no longer be necessary and was now most likely causing clipping of the useful part of the message. To be tested on the physical device!